backend/bot/requirements.txt
python-telegram-bot==20.3
redis==4.6.0
python-dotenv==1.0.0